requirements.txt
mysql-connector-python==8.3.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
python-dotenv==1.0.1
